Добавил эндпоинты для бота

Approved-by: zotovvnsber
Approved-by: Helen B
Approved-by: Иван Яковлев

+++ CarWashService/src/main/java/ru/edu/entity/TimeTableID.java
@@ -18,7 +18,7 @@ public class TimeTableID implements Serializable {
 
     private Instant dateTable;
 
-    private Integer idCarWash;
+    private Long idCarWash;
 
     @Override
     public boolean equals(Object o) {
